+++ hpOneView/resources/settings/appliance_device_snmp_v1_trap_destinations.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 ###
-# (C) Copyright [2019] Hewlett Packard Enterprise Development LP
+# (C) Copyright [2020] Hewlett Packard Enterprise Development LP
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -19,50 +19,89 @@ from __future__ import absolute_import
 from __future__ import division
 from __future__ import print_function
 from __future__ import unicode_literals
-
 from future import standard_library
 
 standard_library.install_aliases()
 
-from hpOneView.resources.resource import ResourceClient
+
+from hpOneView.resources.resource import Resource
+from hpOneView.exceptions import HPOneViewException
 
 
-class ApplianceDeviceSNMPv1TrapDestinations(object):
+class ApplianceDeviceSNMPv1TrapDestinations(Resource):
     """
     ApplianceDeviceSNMPv1TrapDestinations API client.
-    The appliance has the ability to forward events received from monitored or managed server hardware to the specified destinations as SNMPv1 traps.
+    The appliance has the ability to forward events received from monitored or managed
+    server hardware to the specified destinations as SNMPv1 traps.
     """
     URI = '/rest/appliance/trap-destinations'
 
-    def __init__(self, con):
-        self._client = ResourceClient(con, self.URI)
+    def __init__(self, connection, data=None):
+        super(ApplianceDeviceSNMPv1TrapDestinations, self).__init__(connection, data)
 
-    def create(self, resource, id=None, timeout=-1):
+    def create_validation(self, destination, community_string, uri, timeout=-1):
+        """
+        Validate whether a hostname or ip address is a valid trap destination.
+        If validation fails, it returns an error identifying the problem that occurred.
+
+        Args:
+            destination (str): destination ip address.
+            community_string (str) - community string for the snmpv1 trap
+            uri - uri of the snmpv1 trap destination
+            timeout:
+                Timeout in seconds. Wait for task completion by default. The timeout does not abort the operation
+                in OneView, just stop waiting for its completion.
+
+        Return:
+            Returns error message if destination already exists.
+
+        """
+        data = dict(
+            destination=destination,
+            communityString=community_string,
+            uri=uri
+        )
+        validation_uri = "{}/validation".format(self.URI)
+        return self._helper.create(data, uri=validation_uri, timeout=timeout)
+
+    def create(self, data, id=None, timeout=-1):
         """
         Adds the specified trap forwarding destination.
         The trap destination associated with the specified id will be created if trap destination with that id does not exists.
         The id can only be an integer greater than 0.
 
         Args:
-            resource (dict): Object to create.
+            data (dict): Object to create.
+            id: id of the resource to be created
             timeout:
                 Timeout in seconds. Wait for task completion by default. The timeout does not abort the operation
                 in OneView, just stop waiting for its completion.
 
-        Returns:
+        Return:
             dict: Created resource.
 
         """
         if not id:
             available_id = self.__get_first_available_id()
-            uri = '%s/%s' % (self.URI, str(available_id))
+            uri = '{}/{}'.format(self.URI, available_id)
         else:
-            uri = '%s/%s' % (self.URI, str(id))
-        return self._client.create(resource, uri=uri, timeout=timeout)
+            uri = '{}/{}'.format(self.URI, id)
+
+        try:
+            response = self.create_validation(data['destination'], data['communityString'], uri)
+        except HPOneViewException as e:
+            raise HPOneViewException(e.msg)
+
+        return super(ApplianceDeviceSNMPv1TrapDestinations, self).create(data, uri=uri, timeout=timeout)
 
     def __findFirstMissing(self, array, start, end):
         """
         Find the smallest elements missing in a sorted array.
+
+        Args:
+            array - list if ids
+            start - starting index
+            end - ending index
 
         Returns:
             int: The smallest element missing.
@@ -88,7 +127,7 @@ class ApplianceDeviceSNMPv1TrapDestinations(object):
         Returns:
             int: The first available id
         """
-        traps = self.get_all()
+        traps = self._helper.get_all()
         if traps:
             used_ids = [0]
             for trap in traps:
@@ -98,87 +137,3 @@ class ApplianceDeviceSNMPv1TrapDestinations(object):
             return self.__findFirstMissing(used_ids, 0, len(used_ids) - 1)
         else:
             return 1
-
-    def get(self, id_or_uri):
-        """
-        Returns the SNMPv1 trap forwarding destination with the specified ID, if it exists.
-
-        Args:
-            id_or_uri: ID or URI of SNMPv1 trap destination.
-
-        Returns:
-            dict: Appliance SNMPv1 trap destination.
-        """
-        return self._client.get(id_or_uri)
-
-    def get_all(self, start=0, count=-1, filter='', sort=''):
-        """
-        Retrieves all SNMPv1 trap forwarding destinations.
-
-        Args:
-            start:
-                The first item to return, using 0-based indexing.
-                If not specified, the default is 0 - start with the first available item.
-            count:
-                The number of resources to return. A count of -1 requests all items.
-
-                The actual number of items in the response might differ from the requested
-                count if the sum of start and count exceeds the total number of items.
-            filter (list or str):
-                A general filter/query string to narrow the list of items returned. The
-                default is no filter; all resources are returned.
-            sort:
-                The sort order of the returned data set. By default, the sort order is based
-                on create time with the oldest entry first.
-
-        Returns:
-            list: A list of SNMPv1 Trap Destionations.
-        """
-        return self._client.get_all(start, count, filter=filter, sort=sort)
-
-    def get_by(self, field, value):
-        """
-        Gets all SNMPv1 trap forwarding destinations that match the filter.
-        The search is case-insensitive.
-
-        Args:
-            field: field name to filter
-            value: value to filter
-
-        Returns:
-            list: A list of SNMPv1 Trap Destionations.
-        """
-        return self._client.get_by(field, value)
-
-    def delete(self, id_or_uri, timeout=-1):
-        """
-        Deletes SNMPv1 trap forwarding destination based on {Id}.
-
-        Args:
-            id_or_uri: dict object to delete
-            timeout:
-                Timeout in seconds. Wait for task completion by default. The timeout does not abort the operation
-                in OneView, just stop waiting for its completion.
-
-        Returns:
-            bool: Indicates if the resource was successfully deleted.
-
-        """
-        return self._client.delete(id_or_uri, timeout=timeout)
-
-    def update(self, resource, timeout=-1):
-        """
-        Updates the specified trap forwarding destination.
-        The trap destination associated with the specified id will be updated if a trap destination with that id already exists.
-        The id can only be an integer greater than 0.
-
-        Args:
-            resource: dict object with changes.
-            timeout:
-                Timeout in seconds. Wait for task completion by default. The timeout does not abort the operation
-                in OneView, just stop waiting for its completion.
-
-        Returns:
-            dict: Updated appliance SNMPv1 trap destinations.
-        """
-        return self._client.update(resource, timeout=timeout)