+++ tests/unit/test_oneview_client.py
@@ -882,9 +882,9 @@ class OneViewClientTest(unittest.TestCase):
         self.assertIsInstance(self._oneview.appliance_device_snmp_v1_trap_destinations,
                               ApplianceDeviceSNMPv1TrapDestinations)
 
-    def test_lazy_loading_appliance_device_device_snmp_v1_trap_destinations(self):
+    def test_appliance_device_device_snmp_v1_trap_destinations_client(self):
         appliance_device_snmp_v1_trap_destinations = self._oneview.appliance_device_snmp_v1_trap_destinations
-        self.assertEqual(appliance_device_snmp_v1_trap_destinations, self._oneview.appliance_device_snmp_v1_trap_destinations)
+        self.assertNotEqual(appliance_device_snmp_v1_trap_destinations, self._oneview.appliance_device_snmp_v1_trap_destinations)
 
     def test_appliance_device_device_snmp_v3_trap_destinations_has_right_type(self):
         self.assertIsInstance(self._oneview.appliance_device_snmp_v3_trap_destinations,