+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 Extends support of the SDK to OneView REST API version 1600 (OneView v5.20).
 
 #### Features supported with the current release
+- Appliance SNMPv1 Trap Destinations
 - Certificates Server
 - Enclosures
 - Enclosure Groups